step 4 part 2
