+++ app/components/Guides/CreateATest/NextButton.tsx
@@ -6,11 +6,16 @@ import { copy } from "../../../theme/copy";
 import Button from "../../shared/AppButton";
 
 type Props = {
+  isDisabled?: boolean;
   noMargin?: boolean;
   step: number;
 };
 
-export default function NextButton({ noMargin, step }: Props): JSX.Element {
+export default function NextButton({
+  isDisabled,
+  noMargin,
+  step,
+}: Props): JSX.Element {
   const { push } = useRouter();
 
   const handleClick = (): void => {
@@ -19,7 +24,12 @@ export default function NextButton({ noMargin, step }: Props): JSX.Element {
 
   return (
     <Box align="end" margin={noMargin ? undefined : { top: "xlarge" }}>
-      <Button isLarge label={copy.next} onClick={handleClick} type="primary" />
+      <Button
+        isLarge
+        label={copy.next}
+        onClick={handleClick}
+        type={isDisabled ? "disabled" : "primary"}
+      />
     </Box>
   );
 }
