--- app/pages/guides/create-a-test/4-old.tsx
@@ -1,28 +1,0 @@
-import { Box } from "grommet";
-import Image from "next/image";
-
-import { textProps } from "../../../components/Guides/CreateATest/helpers";
-import Layout from "../../../components/Guides/CreateATest/Layout";
-import NextButton from "../../../components/Guides/CreateATest/NextButton";
-import Text from "../../../components/shared/Text";
-import { copy } from "../../../theme/copy";
-import { edgeSize } from "../../../theme/theme";
-
-const step = 4;
-
-export default function CreateATest4(): JSX.Element {
-  return (
-    <Layout>
-      <Text {...textProps}>{copy.runSelectedCode}</Text>
-      <Text {...textProps} margin={{ vertical: "medium" }}>
-        {copy.runSelectedCode2}
-      </Text>
-      <Box alignSelf="center" width="480px">
-        <Image height={60} src="/guides/highlight-line.png" width={850} />
-        <Box height={edgeSize.xxsmall} />
-        <Image height={132} src="/guides/run-1-line.png" width={958} />
-      </Box>
-      <NextButton step={step} />
-    </Layout>
-  );
-}
